+++ demos/supabase-todolist/shared/src/commonMain/kotlin/com/powersync/demos/App.kt
@@ -26,16 +26,19 @@ import com.powersync.demos.screens.SignUpScreen
 import com.powersync.demos.screens.TodosScreen
 import kotlinx.coroutines.runBlocking
 
-
 @Composable
-fun App(factory: DatabaseDriverFactory, modifier: Modifier = Modifier) {
-    val supabase = remember {
-        SupabaseConnector(
-            powerSyncEndpoint = Config.POWERSYNC_URL,
-            supabaseUrl = Config.SUPABASE_URL,
-            supabaseKey = Config.SUPABASE_ANON_KEY
-        )
-    }
+fun App(
+    factory: DatabaseDriverFactory,
+    modifier: Modifier = Modifier,
+) {
+    val supabase =
+        remember {
+            SupabaseConnector(
+                powerSyncEndpoint = Config.POWERSYNC_URL,
+                supabaseUrl = Config.SUPABASE_URL,
+                supabaseKey = Config.SUPABASE_ANON_KEY,
+            )
+        }
     val db = remember { PowerSyncDatabase(factory, schema) }
     val status by db.currentStatus.asFlow().collectAsState(initial = db.currentStatus)
 
@@ -48,9 +51,10 @@ fun App(factory: DatabaseDriverFactory, modifier: Modifier = Modifier) {
     }
 
     val navController = remember { NavController(Screen.Home) }
-    val authViewModel = remember {
-        AuthViewModel(supabase, db, navController)
-    }
+    val authViewModel =
+        remember {
+            AuthViewModel(supabase, db, navController)
+        }
 
     val authState by authViewModel.authState.collectAsState()
     val currentScreen by navController.currentScreen.collectAsState()
@@ -81,7 +85,7 @@ fun App(factory: DatabaseDriverFactory, modifier: Modifier = Modifier) {
 
     when (currentScreen) {
         is Screen.Home -> {
-            if(authState == AuthState.SignedOut) {
+            if (authState == AuthState.SignedOut) {
                 navController.navigate(Screen.SignIn)
             }
 
@@ -93,14 +97,13 @@ fun App(factory: DatabaseDriverFactory, modifier: Modifier = Modifier) {
             HomeScreen(
                 modifier = modifier.background(MaterialTheme.colors.background),
                 items = items,
-                isConnected = status.connected,
                 onSignOutSelected = { handleSignOut() },
                 inputText = listsInputText,
                 onItemClicked = handleOnItemClicked,
                 onItemDeleteClicked = lists.value::onItemDeleteClicked,
                 onAddItemClicked = lists.value::onAddItemClicked,
                 onInputTextChanged = lists.value::onInputTextChanged,
-                hasSynced = hasSyncedLists
+                syncStatus = status,
             )
         }
 
@@ -113,7 +116,7 @@ fun App(factory: DatabaseDriverFactory, modifier: Modifier = Modifier) {
                 modifier = modifier.background(MaterialTheme.colors.background),
                 navController = navController,
                 items = todoItems,
-                isConnected = status.connected,
+                syncStatus = status,
                 inputText = todosInputText,
                 onItemClicked = todos.value::onItemClicked,
                 onItemDoneChanged = todos.value::onItemDoneChanged,
@@ -133,24 +136,24 @@ fun App(factory: DatabaseDriverFactory, modifier: Modifier = Modifier) {
         }
 
         is Screen.SignIn -> {
-            if(authState == AuthState.SignedIn) {
+            if (authState == AuthState.SignedIn) {
                 navController.navigate(Screen.Home)
             }
 
             SignInScreen(
                 navController,
-                authViewModel
+                authViewModel,
             )
         }
 
         is Screen.SignUp -> {
-            if(authState == AuthState.SignedIn) {
+            if (authState == AuthState.SignedIn) {
                 navController.navigate(Screen.Home)
             }
 
             SignUpScreen(
                 navController,
-                authViewModel
+                authViewModel,
             )
         }
     }